refactor web3constants
