+++ wondrous-app/components/Icons/Klaytn.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+export default function Klaytn() {
+  return (
+    <svg width="14" height="22" viewBox="0 0 568 560" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clipPath="url(#clip0_401_2)">
+        <path d="M296.969 278.568L490.705 470.887C592.521 362.89 592.521 194.246 490.705 86.2498" fill="#4F473B" />
+        <path d="M282.821 291.299L91.9204 480.782L282.821 559.972L473.722 480.782" fill="#4F473B" />
+        <path d="M275.761 271.508L473.75 74.937L291.299 0L110.293 436.976L275.761 271.508Z" fill="#4F473B" />
+        <path
+          d="M6.04417e-05 278.568C-0.141705 350.103 27.1906 418.944 76.3547 470.887L265.838 14.1198"
+          fill="#4F473B"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_401_2">
+          <rect width="567.06" height="560" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
