+++ wondrous-app/pages/login.tsx
@@ -17,7 +17,8 @@ import WalletConnectConnector from 'components/WalletConnectors/WalletConnect';
 import signedMessageIsString from 'services/web3/utils/signedMessageIsString';
 import CoinbaseConnector from 'components/WalletConnectors/Coinbase';
 import { getDiscordUrl } from 'utils';
-import { DISCORD_CONNECT_TYPES, GRAPHQL_ERRORS, SUPPORTED_CHAINS } from 'utils/constants';
+import { DISCORD_CONNECT_TYPES, GRAPHQL_ERRORS } from 'utils/constants';
+import { SUPPORTED_CHAINS } from 'utils/web3Constants';
 import OnboardingHeader from 'components/Onboarding/OnboardingLayout/Header';
 import { Layout, OnboardingTitle } from 'components/Onboarding/OnboardingLayout/styles';
 import { Connectors, MainWrapper } from 'components/Onboarding/styles';
