+++ wondrous-app/services/web3/connectors/walletConnect.ts
@@ -1,4 +1,4 @@
-import { RPC_URLS, SUPPORTED_CHAIN_IDS } from 'utils/constants';
+import { RPC_URLS, SUPPORTED_CHAIN_IDS } from 'utils/web3Constants';
 import { WalletConnectConnector } from '@web3-react/walletconnect-connector';
 
 export const walletConnect = new WalletConnectConnector({
