+++ wondrous-app/services/web3/connectors/walletLink.ts
@@ -1,4 +1,4 @@
-import { RPC_URLS } from 'utils/constants';
+import { RPC_URLS } from 'utils/web3Constants';
 import { WalletLinkConnector } from '@web3-react/walletlink-connector';
 
 export const walletLink = new WalletLinkConnector({
