+++ wondrous-app/utils/constants.tsx
@@ -119,65 +119,6 @@ export const COLUMNS_CONFIGURATION = {
   ASSIGNEE: 'assignee',
 };
 
-// Supported Chains (ETHEREUM, POLYGON)
-const SUPPORTED_CHAINS = {
-  1: 'ETH',
-  137: 'MATIC',
-  1666600000: 'HARMONY',
-  42161: 'ARBITRUM',
-  56: 'BSC',
-  288: 'BOBA',
-  10: 'OPTIMISM',
-  8217: 'KLAYTN',
-  43114: 'AVALANCHE',
-};
-
-export const NATIVE_TOKEN_SYMBOL = {
-  1: 'ETH',
-  4: 'ETH',
-  137: 'MATIC',
-  1666600000: 'ONE',
-  42161: 'AETH',
-  56: 'BNB',
-  288: 'ETH',
-  10: 'OP',
-  8217: 'KLAY',
-  43114: 'AVAX',
-};
-
-if (!process.env.NEXT_PUBLIC_PRODUCTION) {
-  SUPPORTED_CHAINS[5] = 'GOERLI';
-}
-
-export const RPC_URLS: { [chainId: number]: string } = {
-  1: process.env.NEXT_PUBLIC_RPC_URL_ETH,
-  5: process.env.NEXT_PUBLIC_RPC_URL_GOERLI,
-  137: process.env.NEXT_PUBLIC_RPC_URL_MATIC,
-  1666600000: process.env.NEXT_PUBLIC_RPC_URL_HARMONY,
-  42161: process.env.NEXT_PUBLIC_RPC_URL_ARBITRUM,
-  56: process.env.NEXT_PUBLIC_RPC_URL_BSC,
-  288: process.env.NEXT_PUBLIC_RPC_URL_BOBA,
-  10: process.env.NEXT_PUBLIC_RPC_URL_OPTIMISM,
-  43114: process.env.NEXT_PUBLIC_RPC_URL_AVALANCHE,
-};
-
-export const CHAIN_VALUE_TO_GNOSIS_TX_SERVICE_URL = {
-  ethereum: 'https://safe-transaction.mainnet.gnosis.io',
-  goerli: 'https://safe-transaction.goerli.gnosis.io',
-  polygon: 'https://safe-transaction.polygon.gnosis.io',
-  harmony: 'https://multisig.t.hmny.io',
-  arbitrum: 'https://safe-transaction.arbitrum.gnosis.io',
-  bsc: 'https://safe-transaction.bsc.gnosis.io',
-  boba: 'https://safe-transaction.mainnet.boba.network',
-  optimism: 'https://safe-transaction.optimism.gnosis.io',
-  avalanche: 'https://safe-transaction.avalanche.gnosis.io',
-};
-
-export const HARMONY_MULTI_SEND_ADDR = '0x998739BFdAAdde7C933B942a68053933098f9EDa';
-export const HARMONY_SAFE_MASTER_COPY = '0x69f4D1788e39c87893C980c06EdF4b7f686e2938';
-export const HARMONY_SAFE_MASTER_COPY2 = '0xfb1bffC9d739B8D520DaF37dF666da4C687191EA';
-export const HARMONY_PROXY_FACTORY = '0xC22834581EbC8527d974F8a1c97E1bEA4EF910BC';
-
 export const DISCORD_CONNECT_TYPES = {
   signup: 'signup',
   login: 'login',
@@ -190,94 +131,6 @@ export const DISCORD_CONNECT_TYPES = {
 export const NOTION_CONNECT_TYPES = {
   TASK_IMPORT: 'taskImport',
 };
-
-export const DEFAULT_ERC20_GAS_LIMIT = '0x3D090'; // TODO hackey == 250000
-
-export const SUPPORTED_CHAIN_IDS = Object.keys(SUPPORTED_CHAINS).map((chainId) => parseInt(chainId, 10));
-
-export const CHAIN_IDS = {
-  ETH: 1,
-  MATIC: 137,
-  GOERLI: 5,
-  HARMONY: 1666600000,
-  ARBITRUM: 42161,
-  BSC: 56,
-  BOBA: 288,
-  OPTIMISM: 10,
-  KLAYTN: 8217,
-  AVALANCHE: 43114,
-};
-
-export const CHAIN_TO_CHAIN_DIPLAY_NAME = {
-  ethereum: 'Ethereum Mainnet',
-  goerli: 'Goerli Testnet',
-  polygon: 'Polygon Mainnet',
-  harmony: 'Harmony Mainnet',
-  arbitrum: 'Arbitrum One',
-  bsc: 'BNB smart chain',
-  boba: 'Boba Mainnet',
-  optimism: 'Optimism Mainnet',
-  klaytn: 'Klaytn Mainnet',
-  avalanche: 'Avalanche',
-};
-
-export const SUPPORTED_CURRENCIES = [
-  {
-    symbol: 'ETH',
-    chains: [1, 4, 288],
-  },
-  {
-    symbol: 'MATIC',
-    chains: [137],
-  },
-  {
-    symbol: 'ONE',
-    chains: [1666600000],
-  },
-  {
-    symbol: 'AETH',
-    chains: [42161],
-  },
-  {
-    symbol: 'BNB',
-    chains: [56],
-  },
-  {
-    symbol: 'OP',
-    chains: [10],
-  },
-  {
-    symbol: 'KLAY',
-    chains: [8217],
-  },
-  {
-    symbol: 'WONDER',
-    chains: [1, 137, 1666600000, 42161, 56, 288, 10, 8127],
-    contracts: {
-      1: '',
-      137: '',
-      1666600000: '',
-      42161: '',
-      56: '',
-      288: '',
-      10: '',
-      8127: '',
-    },
-  },
-  {
-    symbol: 'USDC',
-    chains: [1, 137, 1666600000, 42161, 288, 10],
-    contracts: {
-      1: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
-      137: '0x2791bca1f2de4661ed88a30c99a7a9449aa84174',
-      1666600000: '0x44cED87b9F1492Bf2DCf5c16004832569f7f6cBa',
-      42161: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
-      56: '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d',
-      288: '0x66a2A913e447d6b4BF33EFbec43aAeF87890FBbc',
-      10: '0x7F5c764cBc14f9669B88837ca1490cCa17c31607',
-    },
-  },
-];
 
 export const SIDEBAR_WIDTH = '84px';
 export const SIDEBAR_WIDTH_WITH_DAO = '344px';
@@ -472,7 +325,6 @@ export const filteredColorOptions = Object.keys(POD_COLOR).map((key) => ({
   label: POD_COLOR[key],
   value: key,
 }));
-export { SUPPORTED_CHAINS };
 export const EMAIL_REGEX =
   /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 export const URL_REGEX =
